payment/backend/requirements.txt
dateparser==1.1.7
dnspython==2.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
jsonschema==4.17.3
mysql-connector-python==8.0.31
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.1
requests-file==1.5.1
watchdog==2.3.1
Werkzeug==2.2.2
pika==1.2.0
